requirements.txt
matrix-nio[sso]
python-dotenv
mcp[cli]
openai
mistralai
markdown
